+++ library/src/main/java/com/evernote/android/job/Job.java
@@ -511,7 +511,7 @@ public abstract class Job {
         }
 
         /**
-         * Returns the time the job did run the last time. This is only useful for periodic jobs
+         * Returns the time the job did run the last time. This is only useful for periodic jobs, daily jobs
          * or jobs which were rescheduled. If the job didn't run, yet, then it returns 0.
          *
          * @return The last time the rescheduled or periodic job did run.