+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.2.2
+## 1.2.2 (2018-01-13)
 * Use only the `support-compat` instead of the full `support-v4` library, see #326
 * Use a `ReadWriteLock` for synchronizing the database reads and writes, see #344
 * Update the last run time for daily jobs, see #332